requirements.txt
Flask
Werkzeug
requests
PyJWT
gunicorn
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
twilio
pytest
